enhance the custom loader to handle XAR
We only used the custom loader to load python extension modules but use dlopen for the shared libraries needed by the extension modules. This causes issue if the non-extension shared libraries uses CPython APIs. There are a few cases why this can happen. E.g., if an extension is created using boost.python, the extension module will depends on libboost_python.so . The latter need access CPython APIs.

This diff maintains a list of non-extension libraries that should be loaded by the custom loader. In long term we should enhance the custom loader to load all shared libraries with it.

Differential Revision: [D32632556](https://our.internmc.facebook.com/intern/diff/D32632556/)

ghstack-source-id: 145182239
Pull Request resolved: https://github.com/pytorch/pytorch/pull/69652