[not for land] dbr quant: prototyping script execution order annotation
Summary:

Not for land or review yet, just prototyping and saving work.

This PR implements a simple TorchScript pass to annotate `aten::add`
nodes in the TorchScript graph with op_idx, trace through the model
with an example input, and extract the mapping of op_idx to execution
order.

This may be useful later if we decide to pursue a bridge to
apply the results of DBR quant on a model transformed with
`torch.jit.script`. As a result of the pass introduced in this PR,
we now have a mapping of nodes in the AST to their execution order
when traced with an example input. This information should be enough
to know how to transfer DBR quant state to a scripted model and do
the quantization syntax transforms correctly. This requires some
assumptions which are detailed below (#3 and #4) which we will need
to confirm as well.

The high level things missing for an e2e working prototype are:

1. current PR just recovers the node ID to execution order, a future PR
  will have to use this mapping to transform the graph to a quantized one.
  This should be medium/high eng effort if we brute force it, but doable.
  There may be some tricks we can use here as well (TBD).
2. current PR just handles `aten::add`, will need to expand to cover all quantizable ops.
  We will also need to think through how we handle custom modules and
  functions.
3. current PR assumes that module hierarchy remains the same after `torch.jit.script`,
  that assumption will need to be checked.
4. current PR assumes that we know the mapping from python module/op to
  TorchScript module/op in advance. It doesn't have to be one-to-one,
  but it needs to be known ahead of time. This assumption will need to
  be checked.

Test Plan:

```
python test/test_quantization.py TestQuantizeDBRScript.test_annotate_with_order_logger
```

Reviewers:

Subscribers:

Tasks:

Tags:

ghstack-source-id: ba6aa89cfbefd601155f43762d5ff2f37ce95e53
Pull Request resolved: https://github.com/pytorch/pytorch/pull/69850