Expand composite compliance tests
This PR expands the composite compliance to catch Tensor-subclass-incompatible
operations.

Certain in-place operations can eliminate the typing of the Tensor
subclass. For example:
>>> torch.zeros(input.sizes(), grad.options()).diag().copy_(input)

We want operator writers to either:
(1) Avoid writing code like this
(2) Put code like this into a fast-path and guard the common case (the
slow path) with areAnyTensorSubclassLike checks.

To test for this, we run composite compliance without python mode.

Test Plan:
- run tests

ghstack-source-id: 9a304e2c7b8d31302194bd097bea93844e655ba2
Pull Request resolved: https://github.com/pytorch/pytorch/pull/69990