[quant][core][improvement] Removed preprocessor directives for fbgemm, qnnpack, mkldnn, qengine checks in qconv_unpack and removed extraneous cloning and reassignments
Summary:
The preprocessor directives in qconv_unpack.cpp appear to be extraneous and serve no purpose and have been removed.
The qengine checks also appears to be unnecessary and have been removed.
squeeze_ is an inplace operator and a reassignment is unnecessary and has been removed.
Cloning the weight tensor also appears to be unnecessary since weight is a local tensor and we can directly operate on it -- actually the cloning may still be needed because local weight is a reference to the packed parameters weight? Let's see if this breaks any unit testing

ghstack-source-id: 0abf3d2e9c0338b179ffe902d9bf896753035ede
Pull Request resolved: https://github.com/pytorch/pytorch/pull/74913