[SR][easy] Delete StaticRuntime move ctors
The move constructor is broken because `BlockRunner` stores a reference to `values_`. When `StaticRuntime` is moved constructed, pointer to the root block is moved, but the reference is not updated, leading to confusing runtime errors.

The move assignment operator is broken for the same reason.

We can always fix this by heap-allocating `values_`, but I figured it's fine to just delete it since we typically work with pointers to runtimes (note for reviewers: if it's convenient to have a move ctor I can change this diff to make that change instead)

Differential Revision: [D35224396](https://our.internmc.facebook.com/intern/diff/D35224396/)

ghstack-source-id: 152495991
Pull Request resolved: https://github.com/pytorch/pytorch/pull/74915