[quant][gpu][core] Added quantized linear operator in cudnn
Summary:
This PR is similar to https://github.com/pytorch/pytorch/pull/70622, but for the linear operator.
Unlke PR 70622, this implementations directly uses packed parameters, rather than a refactorization, as was done for the conv operator,
and also directly implements bias & relu.
Currently, int8 matrix multiplication is not supported in cudnn. The ETA for this support is in the first half of April 2022. As
a temporary workaround, we cast our int8 tensors to fp32 prior to matmul.

Test plan:
```
python test/test_quantization.py TestQuantizedLinear.test_qlinear_cudnn
```

ghstack-source-id: 0a2c6c17a50033bdafd6a48ac927488411009158
Pull Request resolved: https://github.com/pytorch/pytorch/pull/73959