[SR] Implement a StaticRuntime::clone() method
Pull Request resolved: https://github.com/pytorch/pytorch/pull/74928

Implement a method for cloning static runtime instances. The motivation for this feature is the new memory planning algorithm. It's too expensive to compute the offsets on each first iteration - we want to use the offsets computed from the first warm up run to instantiate all future `StaticRuntime` instances. This can be achieved by `clone()`-ing a warm prototype instance.
ghstack-source-id: 152734269

Differential Revision: [D35219248](https://our.internmc.facebook.com/intern/diff/D35219248/)