[shard] add ShardedTensor.cpu()
Pull Request resolved: https://github.com/pytorch/pytorch/pull/74941

Add ShardedTensor.cpu() API, we choose to add this first instead of `ShardedTensor.to()` bc the latter one is ambiguous and tricker (i.e. we need to figure out what's the meaning of ShardedTensor.to("cuda")).

Note that when moving ShardedTensor to CPU from GPU, user need to explicitly pass in a new process_group that is compatible with the device, otherwise we will error out.
ghstack-source-id: 152842150

Differential Revision: [D35238133](https://our.internmc.facebook.com/intern/diff/D35238133/)